Add otlp_exporter.cc file
